Moved request validators classes, no sit alongside route validators
* Moved request validators classes, no sit alongside route validators.
* Updated CHANGELOG with most recent changes.
