+++ app/Http/Parameters/Request/Validators/RequestErrorLog.php
@@ -1,21 +1,22 @@
 <?php
 
-namespace App\Validators;
+namespace App\Http\Parameters\Request\Validators;
 
-use App\Validators\Validator as BaseValidator;
+use Validator as BaseValidator;
 use Illuminate\Contracts\Validation\Validator;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\Validator as ValidatorFacade;
 
 /**
- * Validation helper class for resource types, returns the generated validator objects
+ * Validation helper class for request errors, returns the generated validator
+ * object
  *
  * @author Dean Blackborough <dean@g3d-development.com>
  * @copyright Dean Blackborough 2018
  * @license https://github.com/costs-to-expect/api/blob/master/LICENSE
  */
-class ResourceType extends BaseValidator
+class RequestErrorLog extends BaseValidator
 {
     /**
      * Return the validator object for the create request
@@ -28,8 +29,8 @@ class ResourceType extends BaseValidator
     {
         return ValidatorFacade::make(
             $request->all(),
-            Config::get('api.routes.resource_type.validation.POST.fields'),
-            Config::get('api.routes.resource_type.validation.POST.messages')
+            Config::get('api.routes.request.validation.POST.fields'),
+            Config::get('api.routes.request.validation.POST.messages')
         );
     }
 }
