+++ app/Http/Parameters/Request/Validators/ResourceType.php
@@ -1,21 +1,21 @@
 <?php
 
-namespace App\Validators;
+namespace App\Http\Parameters\Request\Validators;
 
-use App\Validators\Validator as BaseValidator;
+use Validator as BaseValidator;
 use Illuminate\Contracts\Validation\Validator;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\Validator as ValidatorFacade;
 
 /**
- * Validation helper class for item category, returns the generated validator objects
+ * Validation helper class for resource types, returns the generated validator objects
  *
  * @author Dean Blackborough <dean@g3d-development.com>
  * @copyright Dean Blackborough 2018
  * @license https://github.com/costs-to-expect/api/blob/master/LICENSE
  */
-class ItemCategory extends BaseValidator
+class ResourceType extends BaseValidator
 {
     /**
      * Return the validator object for the create request
@@ -26,16 +26,10 @@ class ItemCategory extends BaseValidator
      */
     public function create(Request $request): Validator
     {
-        $decode = $this->hash->category()->decode($request->input('category_id'));
-        $category_id = null;
-        if (count($decode) === 1) {
-            $category_id = $decode[0];
-        }
-
         return ValidatorFacade::make(
-            ['category_id' => $category_id],
-            Config::get('api.routes.item_category.validation.POST.fields'),
-            Config::get('api.routes.item_category.validation.POST.messages')
+            $request->all(),
+            Config::get('api.routes.resource_type.validation.POST.fields'),
+            Config::get('api.routes.resource_type.validation.POST.messages')
         );
     }
 }
